Add files via upload
second times the charm. i am not good at this. lol
objects for admins were being controlled by one bool and was causing every object to turn on, but i did not account for a SetActive to read the bool for isAdmin as setting everything in both of those arrays as setting active to true. x.x

Fixed by using the isAdmin bool equaling true and setting 2 other bools to true and false respectively and using those values on some foreach lines that use the bool from those 2 values to either setactive(true) objects that need to turn on and setactive(false) objects that needed to be off.
yay!