Update! June 16th 2025
Wow, been awhile since i've updated the fork of this, but this script now has more gems you can give people for flair, and I'm testing a microphone that can only be grabbed by admins that will make your audio global. That script isn't added here, but I could add it in the future if so.